# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Consultas Médicas",
        "description": "Documentação da API do Sistema de Consultas Médicas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/especialidades": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Lista todas as especialidades",
                "description": "Retorna todas as especialidades cadastradas.",
                "operationId": "c389d24dc22ce3a99228dbcd590d513c",
                "responses": {
                    "200": {
                        "description": "Lista de especialidades retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "id": 1,
                                            "nome": "Clínico Geral",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "nome": "Ortopedia",
                                            "created_at": "2026-07-22T23:07:45.000000Z",
                                            "updated_at": "2026-07-22T23:07:45.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Cadastra uma nova especialidade",
                "description": "Cria uma nova especialidade no sistema.",
                "operationId": "a39fe81ffd83a11e0b34bbe9e35aa56a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Especialidade cadastrada com sucesso."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            }
        },
        "/api/especialidades/{id}": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Busca uma especialidade pelo ID",
                "description": "Retorna uma especialidade específica.",
                "operationId": "1006f0acec16580644cd8237119a0d7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da especialidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialidade encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": {
                                        "id": 1,
                                        "nome": "Clínico Geral",
                                        "created_at": "2026-07-22T23:07:45.000000Z",
                                        "updated_at": "2026-07-22T23:07:45.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Especialidade não encontrada."
                    }
                }
            },
            "put": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Atualiza uma especialidade",
                "description": "Atualiza os dados de uma especialidade existente.",
                "operationId": "a7287b40e277736d8f6515252a9cad2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da especialidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Neurologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Especialidade atualizada com sucesso."
                    },
                    "404": {
                        "description": "Especialidade não encontrada."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Remove uma especialidade",
                "description": "Exclui uma especialidade pelo ID.",
                "operationId": "cb845875c4f08909d727c01c23bb88c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da especialidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialidade removida com sucesso."
                    },
                    "404": {
                        "description": "Especialidade não encontrada."
                    }
                }
            }
        },
        "/api/especialidades/search": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Pesquisa especialidades",
                "description": "Busca especialidades pelo termo informado.",
                "operationId": "4fca131b81278e7276502380ca5fc885",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Texto para pesquisa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Cardio"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da pesquisa."
                    }
                }
            }
        },
...
    },
...
}